fix: moved heroItems to data.tsx
