--- libs/templates/src/lib/home/data.ts
@@ -1,28 +1,0 @@
-import { StatCardContent } from '@deriv-com/components';
-
-export const homeStatData: StatCardContent[] = [
-  {
-    id: 0,
-    color: 'dark',
-    header: '$48M+',
-    description: 'Monthly withdrawals',
-  },
-  {
-    id: 1,
-    color: 'white',
-    header: '$650B+',
-    description: 'Monthly volume',
-  },
-  {
-    id: 2,
-    color: 'dark',
-    header: '168M+',
-    description: 'Monthly deals',
-  },
-  {
-    id: 3,
-    color: 'coral',
-    header: '300K+',
-    description: 'Monthly active traders',
-  },
-];
